Split package tests
