+++ tests/package/test_compression.py
@@ -1,0 +1,44 @@
+from frictionless import Package, helpers
+
+
+IS_UNIX = not helpers.is_platform("windows")
+BASEURL = "https://raw.githubusercontent.com/frictionlessdata/frictionless-py/master/%s"
+
+
+# Compression
+
+
+def test_package_compression_implicit_gz():
+    package = Package("data/compression/datapackage.json")
+    resource = package.get_resource("implicit-gz")
+    assert resource.read_rows() == [
+        {"id": 1, "name": "english"},
+        {"id": 2, "name": "中国人"},
+    ]
+
+
+def test_package_compression_implicit_zip():
+    package = Package("data/compression/datapackage.json")
+    resource = package.get_resource("implicit-zip")
+    assert resource.read_rows() == [
+        {"id": 1, "name": "english"},
+        {"id": 2, "name": "中国人"},
+    ]
+
+
+def test_package_compression_explicit_gz():
+    package = Package("data/compression/datapackage.json")
+    resource = package.get_resource("explicit-gz")
+    assert resource.read_rows() == [
+        {"id": 1, "name": "english"},
+        {"id": 2, "name": "中国人"},
+    ]
+
+
+def test_package_compression_explicit_zip():
+    package = Package("data/compression/datapackage.json")
+    resource = package.get_resource("explicit-zip")
+    assert resource.read_rows() == [
+        {"id": 1, "name": "english"},
+        {"id": 2, "name": "中国人"},
+    ]
